restify-backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jsonfield==3.1.0
phonenumberslite==8.12.44
Pillow==9.0.1
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
swapper==1.3.0
